Caching

+++ app/src/main/java/com/danilovfa/pokemontracker/domain/model/PokemonDetails.kt
@@ -6,6 +6,6 @@ data class PokemonDetails(
     val name: String,
     val sprite: Bitmap,
     val types: List<String>,
-    val weightInKg: Double,
-    val heightInCm: Double
+    val weightInKg: Int,
+    val heightInCm: Int
 )
